Update charts for file downloads; code splitting
